requirements.txt
openai>=1.30.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
feedparser>=6.0.0
